Add parallelism to some critical regions

+++ constr_code/src/grid.hpp
@@ -246,13 +246,13 @@ public:
             boxlen = dx*size;
 
         MyFloat Mean1=0, Mean2=0, Mean3=0;
-        long idx;
+        size_t idx;
 
         for(int ix=0;ix<size;ix++) {
             for(int iy=0;iy<size;iy++) {
                 for(int iz=0;iz<size;iz++) {
 
-                    idx = (long)(ix*size+iy)*size+iz;
+                    idx = static_cast<size_t>((ix*size+iy)*size+iz);
 
                     // position in physical coordinates
                     Pos1[idx]+= ix*dx+dx/2+x0;
